requirements.txt
mcp
bilibili-api-python
aiohttp
brotlicffi
requests
pydantic
tabulate
flask
